Complete los paquetes Nota, Figura y cree el paquete Constantes. Ahora el test corre pero nose usar el JUnit. faltarian las pruebas unitarias.

+++ AlgoHero/src/CancionHero/ElementoDeCompas.java
@@ -3,5 +3,5 @@ package CancionHero;
 public abstract class ElementoDeCompas {
 
 	// devuelve la duracion medida en negras.
-	public abstract float getDuracion();
+	public abstract double getDuracion();
 }
